requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
pandas==2.1.4
numpy==1.26.4